feat: redesign add new category ui
